Standardize numeric field types in the summary models
The summaries and all-time models used f32 for total_seconds and
percent and u8 for the clock minutes/seconds components, while the
stats models already used f64 and u32 for the same concepts. All
fractional durations and percentages are now f64, all clock components
and percent_calculated u32. Breaking, but folds into the not yet
published 0.2.0.

+++ src/model/all_times_since_today.rs
@@ -11,14 +11,14 @@ pub struct AllTimeSinceToday {
     /// `true` if the stats are up to date; when `false`, a 202 response code is returned and stats will be refreshed soon.
     pub is_up_to_date: bool,
     /// A number between 0 and 100 where 100 means the stats are up to date including Today’s time.
-    pub percent_calculated: Option<u8>,
+    pub percent_calculated: Option<u32>,
     pub range: AllTimeSinceTodayRange,
     /// Total time logged since account created as human readable string.
     pub text: String,
     /// Keystroke timeout setting in minutes.
     pub timeout: u32,
     /// Total number of seconds logged since account created.
-    pub total_seconds: f32,
+    pub total_seconds: f64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
